+++ NEWS.md
@@ -1,9 +1,9 @@
-# nanonext 0.2.0.9003 (development)
+# nanonext 0.2.0.9004 (development)
 
 #### New Features
 
 * Aio fields `$result` for a 'sendAio', `$raw` and `$data` for a 'recvAio' may be queried directly, returning their values or else an NA 'unresolved value' if the Aio operation is yet to complete.
-* `is_resolved()` added as an auxiliary function to query whether an Aio has resolved in a non-blocking fashion.
+* `unresolved()` added as an auxiliary function to query whether an Aio is still unresolved in a non-blocking fashion.
 * `is_nul_byte()` added as a helper function for request/reply setups.
 * `survey_time()` added as a convenience function for surveyor/respondent patterns.
 * `logging()` function to specify a global package logging level - currently supports 'error' and 'info'.